Fix Issue #2
The first attempt to create an invoice in FreeAgent requires the
successful negotiation of a new access token between FurAgent and
FreeAgent. Whilst this negotiation has been completing successfully, the
token wasn't been successfully passed back to the caller on the very
first request (after which the old token is read back from the
environment), which caused the first attempt to create a new invoice to
fail. This commit fixes this issue, ensuring that the access token is
correctly returned from FreeAgent.refresh_access_token to the caller.